dynamicpv: Evaluates Present Values and Health Economic Models with Dynamic Pricing and Uptake

The goal of 'dynamicpv' is to provide a simple way to calculate (net) present values and outputs from health economic models (especially cost-effectiveness and budget impact) in discrete time that reflect dynamic pricing and dynamic uptake. Dynamic pricing is also known as life cycle pricing; dynamic uptake is also known as multiple or stacked cohorts, or dynamic disease prevalence. Shafrin (2024) doi:10.1515/fhep-2024-0014 provides an explanation of dynamic value elements, in the context of Generalized Cost Effectiveness Analysis, and Puls (2024) doi:10.1016/j.jval.2024.03.006 reviews challenges of incorporating such dynamic value elements. This package aims to reduce those challenges.

See also

Author

Maintainer: Dominic Muston dominic.muston@msd.com (ORCID)

Other contributors:

John Blischak (ORCID) [contributor]

Merck & Co., Inc., Rahway, NJ, USA and its affiliates [copyright holder, funder]