Calculate present value for a payoff with dynamic pricing and dynamic uptake

Present value of a payoff affected by dynamic pricing, with uptake across multiple cohorts (dynamic uptake)

dynpv(
  uptakes = 1,
  payoffs,
  horizon = length(payoffs),
  tzero = 0,
  prices = rep(1, length(payoffs) + tzero),
  discrate = 0
)

Arguments

uptakes

Vector of patient uptake over time

payoffs

Vector of payoffs of interest (numeric vector)

horizon

Time horizon for the calculation (length must be less than or equal to the length of payoffs)

tzero

Time at the date of calculation, to be used in lookup in prices vector

prices

Vector of price indices through the time horizon of interest

discrate

Discount rate per timestep, corresponding to price index

Value

A tibble of class "dynpv" with the following columns:

j: todo

k: todo

l: todo

t: todo

uj: todo

pk: todo

R: todo

v: todo

pv: todo

Examples

library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

# Obtain dataset
democe <- get_dynfields(
   heemodel = oncpsm,
   payoffs = c("cost_daq_new", "cost_total", "qaly"),
   discount = "disc"
   )

# Obtain short payoff vector of interest
payoffs <- democe |>
   filter(int=="new", model_time<11) |>
   mutate(cost_oth = cost_total - cost_daq_new)
Nt <- nrow(payoffs)

# Example calculation
pv <- dynpv(
   uptakes = rep(1, Nt),
   payoffs = payoffs$cost_oth,
   prices = 1 + (0:(Nt-1))*0.05,
   discrate = 0.08
)
pv
#> # A tibble: 55 × 9
#>        j     k     l     t    uj    pk     R     v    pv
#>    <int> <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1     1     1     0     1     1  355.  1    1      355.
#>  2     1     2     0     2     1  354.  1.05 0.926  344.
#>  3     1     3     0     3     1  354.  1.1  0.857  334.
#>  4     1     4     0     4     1  353.  1.15 0.794  322.
#>  5     1     5     0     5     1  352.  1.2  0.735  310.
#>  6     1     6     0     6     1  351.  1.25 0.681  298.
#>  7     1     7     0     7     1  349.  1.3  0.630  286.
#>  8     1     8     0     8     1  347.  1.35 0.583  274.
#>  9     1     9     0     9     1  345.  1.4  0.540  261.
#> 10     1    10     0    10     1  343.  1.45 0.500  249.
#> # ℹ 45 more rows
summary(pv)
#> Summary of Dynamic Pricing and Uptake
#>      Number of cohorts:             10 
#>      Number of times:               1 
#>      Total uptake:                  10 
#>      Total present value:           15869.53 
#>      Mean present value:            1586.953